smart-asset-manager/requirements.txt
flask
qrcode
python-barcode
pillow